unnecessary (int) cast

+++ core/src/main/java/org/jruby/RubyInteger.java
@@ -19,7 +19,7 @@
  * Copyright (C) 2002-2004 Thomas E Enebo <enebo@acm.org>
  * Copyright (C) 2004 Stefan Matthias Aust <sma@3plus4.de>
  * Copyright (C) 2005 Charles O Nutter <headius@headius.com>
- * 
+ *
  * Alternatively, the contents of this file may be used under the terms of
  * either of the GNU General Public License Version 2 or later (the "GPL"),
  * or the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
@@ -59,7 +59,7 @@ import org.jruby.util.Numeric;
  * @author  jpetersen
  */
 @JRubyClass(name="Integer", parent="Numeric", include="Precision")
-public abstract class RubyInteger extends RubyNumeric { 
+public abstract class RubyInteger extends RubyNumeric {
 
     public static RubyClass createIntegerClass(Ruby runtime) {
         RubyClass integer = runtime.defineClass("Integer", runtime.getNumeric(),
@@ -68,7 +68,7 @@ public abstract class RubyInteger extends RubyNumeric {
 
         integer.index = ClassIndex.INTEGER;
         integer.setReifiedClass(RubyInteger.class);
-        
+
         integer.kindOf = new RubyModule.JavaClassKindOf(RubyInteger.class);
 
         integer.getSingletonClass().undefineMethod("new");
@@ -78,7 +78,7 @@ public abstract class RubyInteger extends RubyNumeric {
         }
 
         integer.defineAnnotatedMethods(RubyInteger.class);
-        
+
         return integer;
     }
 
@@ -89,15 +89,15 @@ public abstract class RubyInteger extends RubyNumeric {
     public RubyInteger(RubyClass rubyClass) {
         super(rubyClass);
     }
-    
+
     public RubyInteger(Ruby runtime, RubyClass rubyClass, boolean useObjectSpace) {
         super(runtime, rubyClass, useObjectSpace);
-    }     
+    }
 
     @Deprecated
     public RubyInteger(Ruby runtime, RubyClass rubyClass, boolean useObjectSpace, boolean canBeTainted) {
         super(runtime, rubyClass, useObjectSpace, canBeTainted);
-    }     
+    }
 
     @Override
     public RubyInteger convertToInteger() {
@@ -111,11 +111,11 @@ public abstract class RubyInteger extends RubyNumeric {
 
     /*  ================
      *  Instance Methods
-     *  ================ 
+     *  ================
      */
 
     /** int_int_p
-     * 
+     *
      */
     @Override
     @JRubyMethod(name = "integer?")
@@ -124,7 +124,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_upto
-     * 
+     *
      */
     @JRubyMethod
     public IRubyObject upto(ThreadContext context, IRubyObject to, Block block) {
@@ -177,7 +177,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_downto
-     * 
+     *
      */
     // TODO: Make callCoerced work in block context...then fix downto, step, and upto.
     @JRubyMethod
@@ -250,7 +250,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_succ
-     * 
+     *
      */
     @JRubyMethod(name = {"succ", "next"})
     public IRubyObject succ(ThreadContext context) {
@@ -279,7 +279,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_chr
-     * 
+     *
      */
     @JRubyMethod(name = "chr", compat = CompatVersion.RUBY1_8)
     public RubyString chr(ThreadContext context) {
@@ -302,7 +302,7 @@ public abstract class RubyInteger extends RubyNumeric {
                 throw runtime.newRangeError(this.toString() + " out of char range");
             } else {
                 if (enc == null) enc = USASCIIEncoding.INSTANCE;
-                return RubyString.newStringNoCopy(runtime, fromEncodedBytes(runtime, enc, (int)value), enc, 0);
+                return RubyString.newStringNoCopy(runtime, fromEncodedBytes(runtime, enc, value), enc, 0);
             }
         }
     }
@@ -332,7 +332,7 @@ public abstract class RubyInteger extends RubyNumeric {
         }
 
         if (n <= 0) throw runtime.newRangeError(this.toString() + " out of char range");
-        
+
         ByteList bytes = new ByteList(n);
         
         try {
@@ -345,7 +345,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_ord
-     * 
+     *
      */
     @JRubyMethod(name = "ord")
     public IRubyObject ord(ThreadContext context) {
@@ -353,7 +353,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** int_to_i
-     * 
+     *
      */
     @JRubyMethod(name = {"to_i", "to_int", "floor", "ceil", "truncate"})
     public IRubyObject to_i() {
@@ -377,14 +377,14 @@ public abstract class RubyInteger extends RubyNumeric {
         if (ndigits > 0) return RubyKernel.new_float(this, this);
         if (ndigits == 0) return this;
         Ruby runtime = context.runtime;
-        
+
         long bytes = (this instanceof RubyFixnum) ? 8 : RubyFixnum.fix2long(callMethod("size"));
         /* If 10**N/2 > this, return 0 */
         /* We have log_256(10) > 0.415241 and log_256(1/2)=-0.125 */
         if (-0.415241 * ndigits - 0.125 > bytes) {
             return RubyFixnum.zero(runtime);
         }
-        
+
         IRubyObject f = Numeric.int_pow(context, 10, -ndigits);
 
         if (this instanceof RubyFixnum && f instanceof RubyFixnum) {
@@ -408,7 +408,7 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** integer_to_r
-     * 
+     *
      */
     @JRubyMethod(name = "to_r", compat = CompatVersion.RUBY1_9)
     public IRubyObject to_r(ThreadContext context) {
@@ -422,7 +422,7 @@ public abstract class RubyInteger extends RubyNumeric {
     public IRubyObject rationalize(ThreadContext context, IRubyObject[] args) {
         return to_r(context);
     }
-    
+
 
     @JRubyMethod(name = "odd?")
     public RubyBoolean odd_p(ThreadContext context) {
@@ -448,25 +448,25 @@ public abstract class RubyInteger extends RubyNumeric {
     }
 
     /** rb_gcd
-     * 
+     *
      */
     @JRubyMethod(name = "gcd", compat = CompatVersion.RUBY1_9)
     public IRubyObject gcd(ThreadContext context, IRubyObject other) {
         checkInteger(context, other);
         return f_gcd(context, this, RubyRational.intValue(context, other));
-    }    
+    }
 
     /** rb_lcm
-     * 
+     *
      */
     @JRubyMethod(name = "lcm", compat = CompatVersion.RUBY1_9)
     public IRubyObject lcm(ThreadContext context, IRubyObject other) {
         checkInteger(context, other);
         return f_lcm(context, this, RubyRational.intValue(context, other));
-    }    
+    }
 
     /** rb_gcdlcm
-     * 
+     *
      */
     @JRubyMethod(name = "gcdlcm", compat = CompatVersion.RUBY1_9)
     public IRubyObject gcdlcm(ThreadContext context, IRubyObject other) {
@@ -489,11 +489,11 @@ public abstract class RubyInteger extends RubyNumeric {
 
     /*  ================
      *  Singleton Methods
-     *  ================ 
+     *  ================
      */
 
     /** rb_int_induced_from
-     * 
+     *
      */
     @JRubyMethod(name = "induced_from", meta = true, compat = CompatVersion.RUBY1_8)
     public static IRubyObject induced_from(ThreadContext context, IRubyObject recv, IRubyObject other) {
